3_anno/Web_Programming/2_progetto/data_migration_django/requirements.txt
asgiref==3.8.1
Django==5.2.1
djangorestframework==3.16.0
psycopg==3.2.9
sqlparse==0.5.3
typing_extensions==4.13.2
